mcp-server-qdrant/pyproject.toml
[project]
name = "mcp-server-qdrant"
version = "0.1.0"
description = "MCP Server for vector database operations using Qdrant"
authors = [
    { name = "XyAgent", email = "agent@xyber.inc" },
]
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "mcp>=1.8.0,<2.0.0",
    "pydantic-settings>=2.0",  
    "uvicorn[standard]",
    "fastmcp>=2.3.4",
    "fastapi>=0.115.12",
    # Qdrant-specific dependencies
    "qdrant-client>=1.7.0",
    "sentence-transformers>=2.2.0",
]

[dependency-groups]
dev = [
    "pytest",
    "httpx", 
    "ruff",
    "black",
    "isort",
]
debug = [
    "debugpy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_server_qdrant"] 

